Add explicit float typecast
Typecast that converts radiusA int into float for division.

+++ src/main/java/com/hcyic1/ball/RubberBall.java
@@ -37,8 +37,8 @@ public class RubberBall extends Ball {
     @Override
     protected Shape makeBall(Point2D center, int radiusA, int radiusB) {
 
-        double x = center.getX() - (radiusA / 2);
-        double y = center.getY() - (radiusB / 2);
+        double x = center.getX() - ((float) radiusA / 2);
+        double y = center.getY() - ((float) radiusB / 2);
 
         return new Ellipse2D.Double(x, y, radiusA, radiusB);
     }
